Move network code from pgx-utils to cargo-pgx
