+++ cargo-pgx/Cargo.toml
@@ -30,6 +30,10 @@ quote = "1.0.18"
 rayon = "1.5.3"
 regex = "1.5.5"
 ureq = { version = "2.4.0", features = ["native-tls"] }
+url = "2.2.2"
+serde = { version = "1.0.137", features = [ "derive" ] }
+serde_derive = "1.0.137"
+serde-xml-rs = "0.5.1"
 syn = { version = "1.0.95", features = [ "extra-traits", "full", "fold", "parsing" ] }
 unescape = "0.1.0"
 fork = "0.1.19"
